Scheduler - Move position calculation methods to the appointment instances
1) Move `getPositionShift` method to the appointment rendering strategies.
2) Move appointment cell position calculation to the appropriate position calculator.
