+++ js/ui/scheduler/appointments/cellPositionCalculator.js
@@ -1,3 +1,6 @@
+import { isDefined } from '../../../core/utils/type';
+import dateUtils from '../../../core/utils/date';
+
 class BaseStrategy {
     constructor(options) {
         this.options = options;
@@ -7,6 +10,11 @@ class BaseStrategy {
     get appointments() { return this.options.dateSettings; } // TODO rename appoitments -> dateSettings
     get viewDataProvider() { return this.options.viewDataProvider; }
     get positionHelper() { return this.options.positionHelper; }
+    get startViewDate() { return this.options.startViewDate; }
+    get viewStartDayHour() { return this.options.viewStartDayHour; }
+    get viewEndDayHour() { return this.options.viewEndDayHour; }
+    get cellDuration() { return this.options.cellDuration; }
+    get getPositionShift() { return this.options.getPositionShiftCallback; }
 
     calculateCellPositions(groupIndices, isAllDayRowAppointment, isRecurrentAppointment) {
         const result = [];
@@ -43,7 +51,7 @@ class BaseStrategy {
             ? appointment.source.groupIndex
             : undefined;
 
-        return this.positionHelper.getCoordinatesByDateInGroup(
+        return this.getCoordinatesByDateInGroup(
             startDate,
             groupIndices,
             isAllDayRowAppointment,
@@ -57,6 +65,113 @@ class BaseStrategy {
             coordinates: position,
             dateSettingIndex
         };
+    }
+
+    getCoordinatesByDate(date, groupIndex, inAllDayRow) {
+        const validGroupIndex = groupIndex || 0;
+
+        const cellInfo = { groupIndex: validGroupIndex, startDate: date, isAllDay: inAllDayRow };
+        const positionByMap = this.viewDataProvider.findCellPositionInMap(cellInfo);
+        if(!positionByMap) {
+            return undefined;
+        }
+
+        const position = this.getCellPosition(
+            positionByMap,
+            inAllDayRow && !this.isVerticalGroupedWorkSpace,
+        );
+
+        const timeShift = inAllDayRow
+            ? 0
+            : this.getTimeShift(date);
+
+        const shift = this.getPositionShift(timeShift, inAllDayRow);
+        const horizontalHMax = this.positionHelper.getHorizontalMax(validGroupIndex, date);
+        const verticalMax = this.positionHelper.getVerticalMax(validGroupIndex);
+
+        return {
+            cellPosition: position.left + shift.cellPosition,
+            top: position.top + shift.top,
+            left: position.left + shift.left,
+            rowIndex: position.rowIndex,
+            columnIndex: position.columnIndex,
+            hMax: horizontalHMax,
+            vMax: verticalMax,
+            groupIndex: validGroupIndex
+        };
+    }
+
+    getCoordinatesByDateInGroup(startDate, groupIndices, inAllDayRow, groupIndex) {
+        const result = [];
+
+        if(this.viewDataProvider.isSkippedDate(startDate)) {
+            return result;
+        }
+
+        let validGroupIndices = [groupIndex];
+
+        if(!isDefined(groupIndex)) {
+            validGroupIndices = this.groupCount
+                ? groupIndices
+                : [0];
+        }
+
+        validGroupIndices.forEach((groupIndex) => {
+            const coordinates = this.getCoordinatesByDate(startDate, groupIndex, inAllDayRow);
+            if(coordinates) {
+                result.push(coordinates);
+            }
+        });
+
+        return result;
+    }
+
+    getCellPosition(cellCoordinates, isAllDayPanel) {
+        const {
+            dateTableCellsMeta,
+            allDayPanelCellsMeta,
+        } = this.DOMMetaData;
+        const {
+            columnIndex,
+            rowIndex,
+        } = cellCoordinates;
+
+        const position = isAllDayPanel
+            ? allDayPanelCellsMeta[columnIndex]
+            : dateTableCellsMeta[rowIndex][columnIndex];
+
+        const validPosition = { ...position };
+
+        if(this.isRtlEnabled) {
+            validPosition.left += position.width;
+        }
+
+        if(validPosition) {
+            validPosition.rowIndex = cellCoordinates.rowIndex;
+            validPosition.columnIndex = cellCoordinates.columnIndex;
+        }
+
+        return validPosition;
+    }
+
+    getTimeShift(date) {
+        const currentDayStart = new Date(date);
+
+        const currentDayEndHour = new Date(new Date(date).setHours(this.viewEndDayHour, 0, 0));
+
+        if(date.getTime() <= currentDayEndHour.getTime()) {
+            currentDayStart.setHours(this.viewStartDayHour, 0, 0, 0);
+        }
+
+        const timeZoneDifference = dateUtils.getTimezonesDifference(date, currentDayStart);
+        const currentDateTime = date.getTime();
+        const currentDayStartTime = currentDayStart.getTime();
+
+        const minTime = this.startViewDate.getTime();
+
+        return (currentDateTime > minTime)
+            ? ((currentDateTime - currentDayStartTime + timeZoneDifference) % this.cellDuration) / this.cellDuration
+            : 0;
     }
 }
 
@@ -83,7 +198,7 @@ class VirtualStrategy extends BaseStrategy {
         const result = [];
 
         dateSettings.forEach(({ source, startDate }, index) => {
-            const coordinate = this.positionHelper.getCoordinatesByDate(
+            const coordinate = this.getCoordinatesByDate(
                 startDate,
                 source.groupIndex,
                 isAllDayRowAppointment
