+++ js/ui/scheduler/appointments/rendering_strategies/strategy_horizontal_month_line.js
@@ -67,6 +67,14 @@ class HorizontalMonthLineRenderingStrategy extends HorizontalAppointmentsStrateg
     needCorrectAppointmentDates() {
         return false;
     }
+
+    getPositionShift() {
+        return {
+            top: 0,
+            left: 0,
+            cellPosition: 0
+        };
+    }
 }
 
 export default HorizontalMonthLineRenderingStrategy;
