+++ js/ui/scheduler/workspaces/ui.scheduler.timeline.js
@@ -248,23 +248,6 @@ class SchedulerTimeline extends SchedulerWorkSpace {
         return 0;
     }
 
-    getPositionShift(timeShift) {
-        const positionShift = super.getPositionShift(timeShift);
-        let left = this.getCellWidth() * timeShift;
-
-        if(this.option('rtlEnabled')) {
-            left *= -1;
-        }
-
-        left += positionShift.left;
-
-        return {
-            top: 0,
-            left: left,
-            cellPosition: left
-        };
-    }
-
     getIntervalDuration(allDay) {
         return this.getCellDuration();
     }
